local-AIDaemon-agentic-rag-adaptative/VERSIONC3/requirements.txt
# ============================================================
# WikiRAG v2.4.0 Dependencies
# ============================================================

# Core
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# NLP
nltk>=3.8.0
rank-bm25>=0.2.2

# Graph
networkx>=3.0

# HTTP/API
requests>=2.28.0
aiohttp>=3.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Optional: Telegram
python-telegram-bot>=20.0

# Development (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
